+++ lib/tbank_invest.dart
@@ -11,7 +11,11 @@ export 'src/invest_http_client.dart';
 export 'src/invest_websocket.dart';
 export 'src/json_types.dart';
 export 'src/models/money_value.dart';
+export 'src/models/instruments_models.dart';
+export 'src/models/orders_models.dart';
 export 'src/models/quotation.dart';
+export 'src/models/sandbox_models.dart';
+export 'src/models/users_models.dart';
 export 'src/tinvest_client.dart';
 
 export 'src/services/instruments_api.dart';